Add exercise 6
I learned the right way to make API calls in a React app + in a
useEffect function - the key fact being that useEffect can't return
anything but the cleanup function. I appreciated seeing how you can do
it both async / await and with a promise style.

I learned about the different ways you can handle exceptions in an API
call - either via a catch or via the second parameter to the promise.
It's fascinating that they do different things - with the former
catching all errors in both the request but also in how you handle the
response and the later catching only errors in the request. That was
really useful to learn.

I learned that setting the "role" property on a div to "alert" tells
screen-readers that this thing is important and will let people know
right away.

I learned that setting something in a state completely clobbers the
existing object. This was *so cool*. I was wondering why you weren't
clearing out error and pokemon in various situations and I'm so glad
you went into this in detail.

I learned that you can define "ErrorBoundaries" around specific
components which acts almost like a try/catch but in a cleaner way.
It's cool you can catch errors in small parts of the app and still
display useful stuff for other areas.

I learned there is a react-error-boundary library we can use that
provides lots of useful features for us - such as explicit fallback,
reset keys, etc.

I learned that error boundaries don't work for things like async calls
or handlers, though - although the react-error-boundary provides some
ways of dealing with this via useErrorHandler().